fix: refactor test structure to handling skipping tests on imports
