+++ tests/comparator/test_array.py
@@ -1,13 +1,9 @@
 import numpy as np
 import pandas as pd
 import polars as pl
-import pytest
-import snowflake.snowpark as sf
 from datacompy.comparator.array import (
     PandasArrayLikeComparator,
     PolarsArrayLikeComparator,
-    SnowflakeArrayLikeComparator,
-    SparkArrayLikeComparator,
 )
 
 
@@ -196,152 +192,4 @@ def test_polars_compare_non_array():
     col2 = pl.Series([{"a": 1}, {"b": 2}])
     comparator = PolarsArrayLikeComparator()
     result = comparator.compare(col1, col2)
-    assert result is None
-
-
-# PySpark
-def test_spark_compare_equal_arrays(spark_session):
-    # Setup
-    data = [([1, 2], [1, 2]), ([3, 4], [3, 4])]
-    df = spark_session.createDataFrame(data, ["col1", "col2"])
-    comparator = SparkArrayLikeComparator()
-
-    # Execute
-    result_col = comparator.compare(dataframe=df, col1="col1", col2="col2")
-    result = df.withColumn("match", result_col)
-
-    # Assert
-    assert result is not None
-    matches = result.select("match").collect()
-    assert all(row.match for row in matches)
-
-
-def test_spark_compare_unequal_arrays(spark_session):
-    # Setup
-    data = [([1, 2], [1, 2]), ([3, 4], [3, 5])]
-    df = spark_session.createDataFrame(data, ["col1", "col2"])
-    comparator = SparkArrayLikeComparator()
-
-    # Execute
-    result_col = comparator.compare(dataframe=df, col1="col1", col2="col2")
-    result = df.withColumn("match", result_col)
-
-    # Assert
-    assert result is not None
-    matches = result.select("match").collect()
-    assert matches[0].match is True
-    assert matches[1].match is False
-
-
-def test_spark_compare_with_nulls(spark_session):
-    # Setup
-    data = [([1, None], [1, None]), ([3, 4], [3, 4])]
-    df = spark_session.createDataFrame(data, ["col1", "col2"])
-    comparator = SparkArrayLikeComparator()
-
-    # Execute
-    result_col = comparator.compare(dataframe=df, col1="col1", col2="col2")
-    result = df.withColumn("match", result_col)
-
-    # Assert
-    assert result is not None
-    matches = result.select("match").collect()
-    assert all(row.match for row in matches)
-
-
-def test_spark_compare_non_array(spark_session):
-    # integers
-    data = [(1, 1), (2, 2)]
-    df = spark_session.createDataFrame(data, ["col1", "col2"])
-    comparator = SparkArrayLikeComparator()
-    result = comparator.compare(dataframe=df, col1="col1", col2="col2")
-    assert result is None
-
-    # floats
-    data = [(1.0, 1.0), (2.0, 2.0)]
-    df = spark_session.createDataFrame(data, ["col1", "col2"])
-    comparator = SparkArrayLikeComparator()
-    result = comparator.compare(dataframe=df, col1="col1", col2="col2")
-    assert result is None
-
-    # dicts
-    data = [({"a": 1}, {"a": 1}), ({"b": 2}, {"b": 2})]
-    df = spark_session.createDataFrame(data, ["col1", "col2"])
-    comparator = SparkArrayLikeComparator()
-    result = comparator.compare(dataframe=df, col1="col1", col2="col2")
-    assert result is None
-
-
-@pytest.mark.snowflake
-def test_snowflake_compare_equal_arrays(snowflake_session):
-    # Setup
-    data = [([1, 2], [1, 2]), ([3, 4], [3, 4])]
-    df = snowflake_session.create_dataframe(data, schema=["col1", "col2"])
-    comparator = SnowflakeArrayLikeComparator()
-
-    # Execute
-    result = comparator.compare(df, "col1", "col2", "match")
-
-    # Assert
-    assert result is not None
-    matches = result.select("match").collect()
-    assert all(row for row in matches)
-
-
-@pytest.mark.snowflake
-def test_snowflake_compare_unequal_arrays(snowflake_session):
-    # Setup
-    data = [([1, 2], [1, 2]), ([3, 4], [3, 5])]
-    df = snowflake_session.create_dataframe(data, schema=["col1", "col2"])
-    comparator = SnowflakeArrayLikeComparator()
-
-    # Execute
-    result = comparator.compare(df, "col1", "col2", "match")
-
-    # Assert
-    assert result is not None
-    matches = result.select("match").collect()
-    assert matches == [
-        sf.Row(match=True),
-        sf.Row(match=False),
-    ]
-
-
-@pytest.mark.snowflake
-def test_snowflake_compare_with_nulls(snowflake_session):
-    # Setup
-    data = [([1, None], [1, None]), ([3, 4], [3, 4])]
-    df = snowflake_session.create_dataframe(data, schema=["col1", "col2"])
-    comparator = SnowflakeArrayLikeComparator()
-
-    # Execute
-    result = comparator.compare(df, "col1", "col2", "match")
-
-    # Assert
-    assert result is not None
-    matches = result.select("match").collect()
-    assert all(row for row in matches)
-
-
-@pytest.mark.snowflake
-def test_snowflake_compare_non_array(snowflake_session):
-    # integers
-    data = [(1, 1), (2, 2)]
-    df = snowflake_session.createDataFrame(data, ["col1", "col2"])
-    comparator = SnowflakeArrayLikeComparator()
-    result = comparator.compare(df, "col1", "col2", "match")
-    assert result is None
-
-    # floats
-    data = [(1.0, 1.0), (2.0, 2.0)]
-    df = snowflake_session.createDataFrame(data, ["col1", "col2"])
-    comparator = SnowflakeArrayLikeComparator()
-    result = comparator.compare(df, "col1", "col2", "match")
-    assert result is None
-
-    # dicts
-    data = [({"a": 1}, {"a": 1}), ({"b": 2}, {"b": 2})]
-    df = snowflake_session.createDataFrame(data, ["col1", "col2"])
-    comparator = SnowflakeArrayLikeComparator()
-    result = comparator.compare(df, "col1", "col2", "match")
     assert result is None
