+++ tests/comparator/test_numeric.py
@@ -1,15 +1,9 @@
 import pandas as pd
 import polars as pl
-import pyspark.sql as ps
-import pytest
-import snowflake.snowpark as sf
 from datacompy.comparator.numeric import (
     PandasNumericComparator,
     PolarsNumericComparator,
-    SnowflakeNumericComparator,
-    SparkNumericComparator,
 )
-from tests.comparator.snowflake_mocks import *  # noqa: F403
 
 
 # tests for PolarsNumericComparator
@@ -119,175 +113,4 @@ def test_pandas_numeric_comparator_error_handling():
     col1 = pd.Series([1.0, 2.0, 3.0])
     col2 = pd.Series([1.0, 2.5, 3.0, 4.0])
     result = comparator.compare(col1, col2)
-    assert result is None
-
-
-# tests for SparkNumericComparator
-def test_spark_numeric_comparator_exact_match(spark_session):
-    comparator = SparkNumericComparator()
-    df = spark_session.createDataFrame(
-        [(1.0, 1.0), (2.0, 2.0), (3.0, 3.0)], ["col1", "col2"]
-    )
-    result_col = comparator.compare(dataframe=df, col1="col1", col2="col2")
-    result = df.withColumn("col_match", result_col)
-    assert result.select(["col_match"]).collect() == [
-        ps.Row(col_match=True),
-        ps.Row(col_match=True),
-        ps.Row(col_match=True),
-    ]
-
-
-def test_spark_numeric_comparator_approximate_match(spark_session):
-    comparator = SparkNumericComparator(rtol=1e-3, atol=1e-3)
-    df = spark_session.createDataFrame(
-        [(1.0, 1.001), (2.0, 2.002), (3.0, 3.003)], ["col1", "col2"]
-    )
-    result_col = comparator.compare(dataframe=df, col1="col1", col2="col2")
-    result = df.withColumn("col_match", result_col)
-    assert result.select(["col_match"]).collect() == [
-        ps.Row(col_match=True),
-        ps.Row(col_match=True),
-        ps.Row(col_match=True),
-    ]
-
-
-def test_spark_numeric_comparator_type_casting(spark_session):
-    comparator = SparkNumericComparator()
-    df = spark_session.createDataFrame([(1, 1.0), (2, 2.0), (3, 3.0)], ["col1", "col2"])
-    result_col = comparator.compare(dataframe=df, col1="col1", col2="col2")
-    result = df.withColumn("col_match", result_col)
-    assert result.select(["col_match"]).collect() == [
-        ps.Row(col_match=True),
-        ps.Row(col_match=True),
-        ps.Row(col_match=True),
-    ]
-
-
-def test_spark_numeric_comparator_nan_handling(spark_session):
-    comparator = SparkNumericComparator()
-    df = spark_session.createDataFrame(
-        [(1.0, 1.0), (float("nan"), float("nan")), (3.0, 3.0)], ["col1", "col2"]
-    )
-    result_col = comparator.compare(dataframe=df, col1="col1", col2="col2")
-    result = df.withColumn("col_match", result_col)
-    assert result.select(["col_match"]).collect() == [
-        ps.Row(col_match=True),
-        ps.Row(col_match=True),
-        ps.Row(col_match=True),
-    ]
-
-
-def test_spark_numeric_comparator_mismatch(spark_session):
-    comparator = SparkNumericComparator()
-    df = spark_session.createDataFrame(
-        [(1.0, 1.0), (2.0, 2.5), (3.0, 3.0)], ["col1", "col2"]
-    )
-    result_col = comparator.compare(dataframe=df, col1="col1", col2="col2")
-    result = df.withColumn("col_match", result_col)
-    assert result.select(["col_match"]).collect() == [
-        ps.Row(col_match=True),
-        ps.Row(col_match=False),
-        ps.Row(col_match=True),
-    ]
-
-
-def test_spark_numeric_comparator_error_handling(spark_session):
-    comparator = SparkNumericComparator()
-    df = spark_session.createDataFrame(
-        [("a", "x"), ("b", "y"), ("c", "z")], ["col1", "col2"]
-    )  # Invalid type for numeric comparison
-    result_col = comparator.compare(dataframe=df, col1="col1", col2="col2")
-    assert result_col is None
-
-
-# tests for SnowflakeNumericComparator
-@pytest.mark.snowflake
-def test_snowflake_numeric_comparator_exact_match(snowflake_session):
-    comparator = SnowflakeNumericComparator()
-    df = snowflake_session.createDataFrame(
-        [(1.0, 1.0), (2.0, 2.0), (3.0, 3.0)], ["col1", "col2"]
-    )
-    result = comparator.compare(
-        dataframe=df, col1="col1", col2="col2", col_match="col_match"
-    )
-    assert result.select(["col_match"]).collect() == [
-        sf.Row(COL_MATCH=True),
-        sf.Row(COL_MATCH=True),
-        sf.Row(COL_MATCH=True),
-    ]
-
-
-@pytest.mark.snowflake
-def test_snowflake_numeric_comparator_approximate_match(snowflake_session):
-    comparator = SnowflakeNumericComparator(rtol=1e-3, atol=1e-3)
-    df = snowflake_session.createDataFrame(
-        [(1.0, 1.001), (2.0, 2.002), (3.0, 3.003)], ["col1", "col2"]
-    )
-    result = comparator.compare(
-        dataframe=df, col1="col1", col2="col2", col_match="col_match"
-    )
-    assert result.select(["col_match"]).collect() == [
-        sf.Row(COL_MATCH=True),
-        sf.Row(COL_MATCH=True),
-        sf.Row(COL_MATCH=True),
-    ]
-
-
-@pytest.mark.snowflake
-def test_snowflake_numeric_comparator_type_casting(snowflake_session):
-    comparator = SnowflakeNumericComparator()
-    df = snowflake_session.createDataFrame(
-        [(1, 1.0), (2, 2.0), (3, 3.0)], ["col1", "col2"]
-    )
-    result = comparator.compare(
-        dataframe=df, col1="col1", col2="col2", col_match="col_match"
-    )
-    assert result.select(["col_match"]).collect() == [
-        sf.Row(COL_MATCH=True),
-        sf.Row(COL_MATCH=True),
-        sf.Row(COL_MATCH=True),
-    ]
-
-
-@pytest.mark.snowflake
-def test_snowflake_numeric_comparator_nan_handling(snowflake_session):
-    comparator = SnowflakeNumericComparator()
-    df = snowflake_session.createDataFrame(
-        [(1.0, 1.0), (float("nan"), float("nan")), (3.0, 3.0)], ["col1", "col2"]
-    )
-    result = comparator.compare(
-        dataframe=df, col1="col1", col2="col2", col_match="col_match"
-    )
-    assert result.select(["col_match"]).collect() == [
-        sf.Row(COL_MATCH=True),
-        sf.Row(COL_MATCH=True),
-        sf.Row(COL_MATCH=True),
-    ]
-
-
-@pytest.mark.snowflake
-def test_snowflake_numeric_comparator_mismatch(snowflake_session):
-    comparator = SnowflakeNumericComparator()
-    df = snowflake_session.createDataFrame(
-        [(1.0, 1.0), (2.0, 2.5), (3.0, 3.0)], ["col1", "col2"]
-    )
-    result = comparator.compare(
-        dataframe=df, col1="col1", col2="col2", col_match="col_match"
-    )
-    assert result.select(["col_match"]).collect() == [
-        sf.Row(COL_MATCH=True),
-        sf.Row(COL_MATCH=False),
-        sf.Row(COL_MATCH=True),
-    ]
-
-
-@pytest.mark.snowflake
-def test_snowflake_numeric_comparator_error_handling(snowflake_session):
-    comparator = SnowflakeNumericComparator()
-    df = snowflake_session.createDataFrame(
-        [("a", "x"), ("b", "y"), ("c", "z")], ["col1", "col2"]
-    )  # Invalid type for numeric comparison
-    result = comparator.compare(
-        dataframe=df, col1="col1", col2="col2", col_match="col_match"
-    )
     assert result is None
