+++ tests/comparator/test_string.py
@@ -1,16 +1,10 @@
 import pandas as pd
 import polars as pl
-import pyspark.sql as ps
-import pytest
-import snowflake.snowpark as sf
 from datacompy.comparator.string import (
     PandasStringComparator,
     PolarsStringComparator,
-    SnowflakeStringComparator,
-    SparkStringComparator,
     pandas_compare_string_and_date_columns,
 )
-from tests.comparator.snowflake_mocks import *  # noqa: F403
 
 
 # tests for PolarsStringComparator
@@ -140,178 +134,6 @@ def test_pandas_string_comparator_error_handling():
     col1 = pd.Series(["x", "y", "z"])
     col2 = pd.Series(["x", "y", "z", "c"])
     result = comparator.compare(col1, col2)
-    assert result is None
-
-
-# tests for SparkStringComparator
-def test_spark_string_comparator_exact_match(spark_session):
-    comparator = SparkStringComparator()
-    df = spark_session.createDataFrame(
-        [("a", "a"), ("b", "b"), ("c", "c")], ["col1", "col2"]
-    )
-    result_col = comparator.compare(dataframe=df, col1="col1", col2="col2")
-    result = df.withColumn("col_match", result_col)
-    assert result.select(["col_match"]).collect() == [
-        ps.Row(col_match=True),
-        ps.Row(col_match=True),
-        ps.Row(col_match=True),
-    ]
-
-
-def test_spark_string_comparator_case_space_insensitivity(spark_session):
-    df = spark_session.createDataFrame(
-        [("a", " a"), ("b", "   B  "), ("c", "C")], ["col1", "col2"]
-    )
-
-    comparator = SparkStringComparator(ignore_case=True, ignore_space=True)
-    result_col = comparator.compare(dataframe=df, col1="col1", col2="col2")
-    result = df.withColumn("col_match", result_col)
-    assert result.select(["col_match"]).collect() == [
-        ps.Row(col_match=True),
-        ps.Row(col_match=True),
-        ps.Row(col_match=True),
-    ]
-
-    comparator = SparkStringComparator(ignore_case=True, ignore_space=False)
-    result_col = comparator.compare(dataframe=df, col1="col1", col2="col2")
-    result = df.withColumn("col_match", result_col)
-    assert result.select(["col_match"]).collect() == [
-        ps.Row(col_match=False),
-        ps.Row(col_match=False),
-        ps.Row(col_match=True),
-    ]
-
-    comparator = SparkStringComparator(ignore_case=False, ignore_space=True)
-    result_col = comparator.compare(dataframe=df, col1="col1", col2="col2")
-    result = df.withColumn("col_match", result_col)
-    assert result.select(["col_match"]).collect() == [
-        ps.Row(col_match=True),
-        ps.Row(col_match=False),
-        ps.Row(col_match=False),
-    ]
-
-    comparator = SparkStringComparator(ignore_case=False, ignore_space=False)
-    result_col = comparator.compare(dataframe=df, col1="col1", col2="col2")
-    result = df.withColumn("col_match", result_col)
-    assert result.select(["col_match"]).collect() == [
-        ps.Row(col_match=False),
-        ps.Row(col_match=False),
-        ps.Row(col_match=False),
-    ]
-
-
-def test_spark_string_comparator_nan_handling(spark_session):
-    comparator = SparkStringComparator()
-    df = spark_session.createDataFrame(
-        [("a", "a"), (float("nan"), float("nan")), ("c", "c")], ["col1", "col2"]
-    )
-    result_col = comparator.compare(dataframe=df, col1="col1", col2="col2")
-    result = df.withColumn("col_match", result_col)
-    assert result.select(["col_match"]).collect() == [
-        ps.Row(col_match=True),
-        ps.Row(col_match=True),
-        ps.Row(col_match=True),
-    ]
-
-
-def test_spark_string_comparator_error_handling(spark_session):
-    comparator = SparkStringComparator()
-    df = spark_session.createDataFrame(
-        [(1, 2), (3, 4), (5, 6)], ["col1", "col2"]
-    )  # Invalid type for string comparison
-    result_col = comparator.compare(dataframe=df, col1="col1", col2="col2")
-    assert result_col is None
-
-
-# tests for SnowflakeStringComparator
-@pytest.mark.snowflake
-def test_snowflake_string_comparator_exact_match(snowflake_session):
-    comparator = SnowflakeStringComparator()
-    df = snowflake_session.createDataFrame(
-        [("a", "a"), ("b", "b"), ("c", "c")], ["col1", "col2"]
-    )
-    result = comparator.compare(
-        dataframe=df, col1="col1", col2="col2", col_match="col_match"
-    )
-    assert result.select(["col_match"]).collect() == [
-        sf.Row(col_match=True),
-        sf.Row(col_match=True),
-        sf.Row(col_match=True),
-    ]
-
-
-@pytest.mark.snowflake
-def test_snowflake_string_comparator_case_space_insensitivity(snowflake_session):
-    df = snowflake_session.createDataFrame(
-        [("a", " a"), ("b", "   B  "), ("c", "C")], ["col1", "col2"]
-    )
-
-    comparator = SnowflakeStringComparator(ignore_case=True, ignore_space=True)
-    result = comparator.compare(
-        dataframe=df, col1="col1", col2="col2", col_match="col_match"
-    )
-    assert result.select(["col_match"]).collect() == [
-        sf.Row(col_match=True),
-        sf.Row(col_match=True),
-        sf.Row(col_match=True),
-    ]
-
-    comparator = SnowflakeStringComparator(ignore_case=True, ignore_space=False)
-    result = comparator.compare(
-        dataframe=df, col1="col1", col2="col2", col_match="col_match"
-    )
-    assert result.select(["col_match"]).collect() == [
-        sf.Row(col_match=False),
-        sf.Row(col_match=False),
-        sf.Row(col_match=True),
-    ]
-
-    comparator = SnowflakeStringComparator(ignore_case=False, ignore_space=True)
-    result = comparator.compare(
-        dataframe=df, col1="col1", col2="col2", col_match="col_match"
-    )
-    assert result.select(["col_match"]).collect() == [
-        sf.Row(col_match=True),
-        sf.Row(col_match=False),
-        sf.Row(col_match=False),
-    ]
-
-    comparator = SnowflakeStringComparator(ignore_case=False, ignore_space=False)
-    result = comparator.compare(
-        dataframe=df, col1="col1", col2="col2", col_match="col_match"
-    )
-    assert result.select(["col_match"]).collect() == [
-        sf.Row(col_match=False),
-        sf.Row(col_match=False),
-        sf.Row(col_match=False),
-    ]
-
-
-@pytest.mark.snowflake
-def test_snowflake_string_comparator_null_handling(snowflake_session):
-    comparator = SnowflakeStringComparator()
-    df = snowflake_session.createDataFrame(
-        [("a", "a"), (None, None), ("c", "c")], ["col1", "col2"]
-    )
-    result = comparator.compare(
-        dataframe=df, col1="col1", col2="col2", col_match="col_match"
-    )
-    assert result.select(["col_match"]).collect() == [
-        sf.Row(col_match=True),
-        sf.Row(col_match=True),
-        sf.Row(col_match=True),
-    ]
-
-
-@pytest.mark.snowflake
-def test_snowflake_string_comparator_error_handling(snowflake_session):
-    comparator = SnowflakeStringComparator()
-    df = snowflake_session.createDataFrame(
-        [(1, 2), (3, 4), (5, 6)], ["col1", "col2"]
-    )  # Invalid type for string comparison
-    result = comparator.compare(
-        dataframe=df, col1="col1", col2="col2", col_match="col_match"
-    )
     assert result is None
