+++ tests/test_snowflake.py
@@ -20,7 +20,6 @@ Testing out the datacompy functionality
 import io
 import logging
 import os
-import sys
 import tempfile
 from datetime import datetime
 from decimal import Decimal
@@ -30,6 +29,9 @@ from unittest import mock
 import numpy as np
 import pandas as pd
 import pytest
+
+pytest.importorskip("snowflake.snowpark")
+
 from datacompy.snowflake import (
     SnowflakeCompare,
     _generate_id_within_group,
@@ -47,13 +49,10 @@ from snowflake.snowpark.types import (
     StructType,
 )
 
-logging.basicConfig(stream=sys.stdout, level=logging.DEBUG)
-
 pd.DataFrame.iteritems = pd.DataFrame.items  # Pandas 2+ compatability
 np.bool = np.bool_  # Numpy 1.24.3+ comptability
 
 
-@pytest.mark.snowflake
 def test_numeric_columns_equal_abs(snowflake_session):
     data = """A|B|EXPECTED
 1|1|True
@@ -69,7 +68,6 @@ NULL|NULL|True"""
     assert_series_equal(expect_out, actual_out, check_names=False)
 
 
-@pytest.mark.snowflake
 def test_numeric_columns_equal_rel(snowflake_session):
     data = """A|B|EXPECTED
 1|1|True
@@ -84,7 +82,6 @@ NULL|NULL|True"""
     assert_series_equal(expect_out, actual_out, check_names=False)
 
 
-@pytest.mark.snowflake
 def test_string_columns_equal(snowflake_session):
     data = """A|B|EXPECTED
 Hi|Hi|True
@@ -106,7 +103,6 @@ something||False
     assert_series_equal(expect_out, actual_out, check_names=False)
 
 
-@pytest.mark.snowflake
 def test_string_columns_equal_with_ignore_spaces(snowflake_session):
     data = """A|B|EXPECTED
 Hi|Hi|True
@@ -130,7 +126,6 @@ something||False
     assert_series_equal(expect_out, actual_out, check_names=False)
 
 
-@pytest.mark.snowflake
 def test_date_columns_equal(snowflake_session):
     data = """A|B|EXPECTED
 2017-01-01|2017-01-01|True
@@ -160,7 +155,6 @@ def test_date_columns_equal(snowflake_session):
     assert_series_equal(expect_out, actual_out_rev, check_names=False)
 
 
-@pytest.mark.snowflake
 def test_date_columns_equal_with_ignore_spaces(snowflake_session):
     data = """A|B|EXPECTED
 2017-01-01|2017-01-01   |True
@@ -198,7 +192,6 @@ def test_date_columns_equal_with_ignore_spaces(snowflake_session):
     assert_series_equal(expect_out, actual_out_rev, check_names=False)
 
 
-@pytest.mark.snowflake
 def test_columns_equal_same_type_dif_length(snowflake_session):
     schema = StructType(
         [
@@ -227,7 +220,6 @@ def test_columns_equal_same_type_dif_length(snowflake_session):
     )
 
 
-@pytest.mark.snowflake
 def test_date_columns_unequal(snowflake_session):
     """I want datetime fields to match with dates stored as strings"""
     data = [{"A": "2017-01-01", "B": "2017-01-02"}, {"A": "2017-01-01"}]
@@ -245,7 +237,6 @@ def test_date_columns_unequal(snowflake_session):
     assert not columns_equal(df, "B", "A_DT", "ACTUAL").toPandas()["ACTUAL"].any()
 
 
-@pytest.mark.snowflake
 def test_bad_date_columns(snowflake_session):
     """If strings can't be coerced into dates then it should be false for the
     whole column.
@@ -261,7 +252,6 @@ def test_bad_date_columns(snowflake_session):
     assert columns_equal(df, "A_DT", "B", "ACTUAL").toPandas()["ACTUAL"].any()
 
 
-@pytest.mark.snowflake
 def test_rounded_date_columns(snowflake_session):
     """If strings can't be coerced into dates then it should be false for the
     whole column.
@@ -280,7 +270,6 @@ def test_rounded_date_columns(snowflake_session):
     assert_series_equal(actual, expected, check_names=False)
 
 
-@pytest.mark.snowflake
 def test_decimal_float_columns_equal(snowflake_session):
     data = [
         {"A": Decimal("1"), "B": 1, "EXPECTED": True},
@@ -299,7 +288,6 @@ def test_decimal_float_columns_equal(snowflake_session):
     assert_series_equal(expect_out, actual_out, check_names=False)
 
 
-@pytest.mark.snowflake
 def test_decimal_float_columns_equal_rel(snowflake_session):
     data = [
         {"A": Decimal("1"), "B": 1, "EXPECTED": True},
@@ -320,7 +308,6 @@ def test_decimal_float_columns_equal_rel(snowflake_session):
     assert_series_equal(expect_out, actual_out, check_names=False)
 
 
-@pytest.mark.snowflake
 def test_decimal_columns_equal(snowflake_session):
     data = [
         {"A": Decimal("1"), "B": Decimal("1"), "EXPECTED": True},
@@ -343,7 +330,6 @@ def test_decimal_columns_equal(snowflake_session):
     assert_series_equal(expect_out, actual_out, check_names=False)
 
 
-@pytest.mark.snowflake
 def test_decimal_columns_equal_rel(snowflake_session):
     data = [
         {"A": Decimal("1"), "B": Decimal("1"), "EXPECTED": True},
@@ -368,7 +354,6 @@ def test_decimal_columns_equal_rel(snowflake_session):
     assert_series_equal(expect_out, actual_out, check_names=False)
 
 
-@pytest.mark.snowflake
 def test_infinity_and_beyond(snowflake_session):
     # https://spark.apache.org/docs/latest/sql-ref-datatypes.html#positivenegative-infinity-semantics
     # Positive/negative infinity multiplied by 0 returns NaN.
@@ -389,7 +374,6 @@ def test_infinity_and_beyond(snowflake_session):
     assert_series_equal(expect_out, actual_out, check_names=False)
 
 
-@pytest.mark.snowflake
 def test_compare_table_setter_bad(snowflake_session):
     # Invalid table name
     with raises(ValueError, match=r"invalid_table_name_1 is not a valid table name."):
@@ -407,7 +391,6 @@ def test_compare_table_setter_bad(snowflake_session):
     "datacompy.snowflake.SnowflakeCompare._validate_dataframe", new=mock.MagicMock()
 )
 @mock.patch("datacompy.snowflake.SnowflakeCompare._compare", new=mock.MagicMock())
-@pytest.mark.snowflake
 def test_compare_table_unique_names(snowflake_session):
     # Assert that two tables with the same name but from a different DB/Schema have unique names
     # Same schema/name, different DB
@@ -429,7 +412,6 @@ def test_compare_table_unique_names(snowflake_session):
     assert compare.df1_name != compare.df2_name
 
 
-@pytest.mark.snowflake
 def test_compare_table_setter_good(snowflake_session):
     data = """ACCT_ID,DOLLAR_AMT,NAME,FLOAT_FLD,DATE_FLD
     10000001234,123.4,George Michael Bluth,14530.155,
@@ -465,7 +447,6 @@ def test_compare_table_setter_good(snowflake_session):
     assert compare.join_columns == ["ACCT_ID"]
 
 
-@pytest.mark.snowflake
 def test_compare_df_setter_bad(snowflake_session):
     pdf = pd.DataFrame([{"A": 1, "C": 2}, {"A": 2, "C": 2}])
     df = snowflake_session.createDataFrame(pdf)
@@ -484,7 +465,6 @@ def test_compare_df_setter_bad(snowflake_session):
     )
 
 
-@pytest.mark.snowflake
 def test_compare_df_setter_good(snowflake_session):
     df1 = snowflake_session.createDataFrame([{"A": 1, "B": 2}, {"A": 2, "B": 2}])
     df2 = snowflake_session.createDataFrame([{"A": 1, "B": 2}, {"A": 2, "B": 3}])
@@ -496,7 +476,6 @@ def test_compare_df_setter_good(snowflake_session):
     assert compare.join_columns == ["A", "B"]
 
 
-@pytest.mark.snowflake
 def test_compare_df_setter_different_cases(snowflake_session):
     df1 = snowflake_session.createDataFrame([{"A": 1, "B": 2}, {"A": 2, "B": 2}])
     df2 = snowflake_session.createDataFrame([{"A": 1, "B": 2}, {"A": 2, "B": 3}])
@@ -504,7 +483,6 @@ def test_compare_df_setter_different_cases(snowflake_session):
     assert compare.df1.toPandas().equals(df1.toPandas())
 
 
-@pytest.mark.snowflake
 def test_columns_overlap(snowflake_session):
     df1 = snowflake_session.createDataFrame([{"A": 1, "B": 2}, {"A": 2, "B": 2}])
     df2 = snowflake_session.createDataFrame([{"A": 1, "B": 2}, {"A": 2, "B": 3}])
@@ -514,7 +492,6 @@ def test_columns_overlap(snowflake_session):
     assert compare.intersect_columns() == {"A", "B"}
 
 
-@pytest.mark.snowflake
 def test_columns_no_overlap(snowflake_session):
     df1 = snowflake_session.createDataFrame(
         [{"A": 1, "B": 2, "C": "HI"}, {"A": 2, "B": 2, "C": "YO"}]
@@ -528,7 +505,6 @@ def test_columns_no_overlap(snowflake_session):
     assert compare.intersect_columns() == {"A", "B"}
 
 
-@pytest.mark.snowflake
 def test_columns_maintain_order_through_set_operations(snowflake_session):
     pdf1 = pd.DataFrame(
         {
@@ -560,7 +536,6 @@ def test_columns_maintain_order_through_set_operations(snowflake_session):
     assert list(compare.intersect_columns()) == ["JOIN", "G", "B", "H", "A"]
 
 
-@pytest.mark.snowflake
 def test_10k_rows(snowflake_session):
     rng = np.random.default_rng()
     pdf = pd.DataFrame(rng.integers(0, 100, size=(10000, 2)), columns=["B", "C"])
@@ -589,7 +564,6 @@ def test_10k_rows(snowflake_session):
     assert not compare_no_tol.intersect_rows_match()
 
 
-@pytest.mark.snowflake
 def test_10k_rows_abs_tol_per_column(snowflake_session):
     rng = np.random.default_rng()
     pdf = pd.DataFrame(rng.integers(0, 100, size=(10000, 2)), columns=["B", "C"])
@@ -611,7 +585,6 @@ def test_10k_rows_abs_tol_per_column(snowflake_session):
     assert compare_tol.intersect_rows_match()
 
 
-@pytest.mark.snowflake
 def test_10k_rows_abs_tol_per_column_default(snowflake_session):
     rng = np.random.default_rng()
     pdf = pd.DataFrame(rng.integers(0, 100, size=(10000, 2)), columns=["B", "C"])
@@ -634,7 +607,6 @@ def test_10k_rows_abs_tol_per_column_default(snowflake_session):
     assert not compare_tol.intersect_rows_match()
 
 
-@pytest.mark.snowflake
 def test_10k_rows_rel_tol_per_column(snowflake_session):
     rng = np.random.default_rng()
     pdf = pd.DataFrame(rng.integers(0, 100, size=(10000, 2)), columns=["B", "C"])
@@ -656,7 +628,6 @@ def test_10k_rows_rel_tol_per_column(snowflake_session):
     assert compare_tol.intersect_rows_match()
 
 
-@pytest.mark.snowflake
 def test_10k_rows_rel_tol_per_column_default(snowflake_session):
     rng = np.random.default_rng()
     pdf = pd.DataFrame(rng.integers(0, 100, size=(10000, 2)), columns=["B", "C"])
@@ -679,7 +650,6 @@ def test_10k_rows_rel_tol_per_column_default(snowflake_session):
     assert not compare_tol.intersect_rows_match()
 
 
-@pytest.mark.snowflake
 def test_subset(snowflake_session, caplog):
     caplog.set_level(logging.DEBUG)
     df1 = snowflake_session.createDataFrame(
@@ -690,7 +660,6 @@ def test_subset(snowflake_session, caplog):
     assert comp.subset()
 
 
-@pytest.mark.snowflake
 def test_not_subset(snowflake_session, caplog):
     caplog.set_level(logging.INFO)
     df1 = snowflake_session.createDataFrame(
@@ -704,7 +673,6 @@ def test_not_subset(snowflake_session, caplog):
     assert "C: 1 / 2 (50.00%) match" in caplog.text
 
 
-@pytest.mark.snowflake
 def test_large_subset(snowflake_session):
     rng = np.random.default_rng()
     pdf = pd.DataFrame(rng.integers(0, 100, size=(10000, 2)), columns=["B", "C"])
@@ -718,7 +686,6 @@ def test_large_subset(snowflake_session):
     assert comp.subset()
 
 
-@pytest.mark.snowflake
 def test_string_joiner(snowflake_session):
     df1 = snowflake_session.createDataFrame([{"AB": 1, "BC": 2}, {"AB": 2, "BC": 2}])
     df2 = snowflake_session.createDataFrame([{"AB": 1, "BC": 2}, {"AB": 2, "BC": 2}])
@@ -726,7 +693,6 @@ def test_string_joiner(snowflake_session):
     assert compare.matches()
 
 
-@pytest.mark.snowflake
 def test_decimal_with_joins(snowflake_session):
     df1 = snowflake_session.createDataFrame(
         [{"A": Decimal("1"), "B": 2}, {"A": Decimal("2"), "B": 2}]
@@ -739,7 +705,6 @@ def test_decimal_with_joins(snowflake_session):
     assert compare.intersect_rows_match()
 
 
-@pytest.mark.snowflake
 def test_decimal_with_nulls(snowflake_session):
     df1 = snowflake_session.createDataFrame(
         [{"A": 1, "B": Decimal("2")}, {"A": 2, "B": Decimal("2")}]
@@ -754,7 +719,6 @@ def test_decimal_with_nulls(snowflake_session):
     assert compare.intersect_rows_match()
 
 
-@pytest.mark.snowflake
 def test_strings_with_joins(snowflake_session):
     df1 = snowflake_session.createDataFrame([{"A": "HI", "B": 2}, {"A": "BYE", "B": 2}])
     df2 = snowflake_session.createDataFrame([{"A": "HI", "B": 2}, {"A": "BYE", "B": 2}])
@@ -765,7 +729,6 @@ def test_strings_with_joins(snowflake_session):
     assert compare.intersect_rows_match()
 
 
-@pytest.mark.snowflake
 def test_temp_column_name(snowflake_session):
     df1 = snowflake_session.createDataFrame([{"A": "HI", "B": 2}, {"A": "BYE", "B": 2}])
     df2 = snowflake_session.createDataFrame(
@@ -775,7 +738,6 @@ def test_temp_column_name(snowflake_session):
     assert actual == "_TEMP_0"
 
 
-@pytest.mark.snowflake
 def test_temp_column_name_one_has(snowflake_session):
     df1 = snowflake_session.createDataFrame(
         [{"_TEMP_0": "HI", "B": 2}, {"_TEMP_0": "BYE", "B": 2}]
@@ -787,7 +749,6 @@ def test_temp_column_name_one_has(snowflake_session):
     assert actual == "_TEMP_1"
 
 
-@pytest.mark.snowflake
 def test_temp_column_name_both_have_temp_1(snowflake_session):
     df1 = snowflake_session.createDataFrame(
         [{"_TEMP_0": "HI", "B": 2}, {"_TEMP_0": "BYE", "B": 2}]
@@ -803,7 +764,6 @@ def test_temp_column_name_both_have_temp_1(snowflake_session):
     assert actual == "_TEMP_1"
 
 
-@pytest.mark.snowflake
 def test_temp_column_name_both_have_temp_2(snowflake_session):
     df1 = snowflake_session.createDataFrame(
         [{"_TEMP_0": "HI", "B": 2}, {"_TEMP_0": "BYE", "B": 2}]
@@ -819,7 +779,6 @@ def test_temp_column_name_both_have_temp_2(snowflake_session):
     assert actual == "_TEMP_2"
 
 
-@pytest.mark.snowflake
 def test_temp_column_name_one_already(snowflake_session):
     df1 = snowflake_session.createDataFrame(
         [{"_TEMP_1": "HI", "B": 2}, {"_TEMP_1": "BYE", "B": 2}]
@@ -838,7 +797,6 @@ def test_temp_column_name_one_already(snowflake_session):
 # Duplicate testing!
 
 
-@pytest.mark.snowflake
 def test_simple_dupes_one_field(snowflake_session):
     df1 = snowflake_session.createDataFrame([{"A": 1, "B": 2}, {"A": 1, "B": 2}])
     df2 = snowflake_session.createDataFrame([{"A": 1, "B": 2}, {"A": 1, "B": 2}])
@@ -848,7 +806,6 @@ def test_simple_dupes_one_field(snowflake_session):
     compare.report()
 
 
-@pytest.mark.snowflake
 def test_simple_dupes_two_fields(snowflake_session):
     df1 = snowflake_session.createDataFrame(
         [{"A": 1, "B": 2}, {"A": 1, "B": 2, "C": 2}]
@@ -862,7 +819,6 @@ def test_simple_dupes_two_fields(snowflake_session):
     compare.report()
 
 
-@pytest.mark.snowflake
 def test_simple_dupes_one_field_two_vals_1(snowflake_session):
     df1 = snowflake_session.createDataFrame([{"A": 1, "B": 2}, {"A": 1, "B": 0}])
     df2 = snowflake_session.createDataFrame([{"A": 1, "B": 2}, {"A": 1, "B": 0}])
@@ -872,7 +828,6 @@ def test_simple_dupes_one_field_two_vals_1(snowflake_session):
     compare.report()
 
 
-@pytest.mark.snowflake
 def test_simple_dupes_one_field_two_vals_2(snowflake_session):
     df1 = snowflake_session.createDataFrame([{"A": 1, "B": 2}, {"A": 1, "B": 0}])
     df2 = snowflake_session.createDataFrame([{"A": 1, "B": 2}, {"A": 2, "B": 0}])
@@ -885,7 +840,6 @@ def test_simple_dupes_one_field_two_vals_2(snowflake_session):
     compare.report()
 
 
-@pytest.mark.snowflake
 def test_simple_dupes_one_field_three_to_two_vals(snowflake_session):
     df1 = snowflake_session.createDataFrame(
         [{"A": 1, "B": 2}, {"A": 1, "B": 0}, {"A": 1, "B": 0}]
@@ -902,7 +856,6 @@ def test_simple_dupes_one_field_three_to_two_vals(snowflake_session):
     assert "(First 2 Columns)" in compare.report(column_count=2)
 
 
-@pytest.mark.snowflake
 def test_dupes_from_real_data(snowflake_session):
     data = """ACCT_ID,ACCT_SFX_NUM,TRXN_POST_DT,TRXN_POST_SEQ_NUM,TRXN_AMT,TRXN_DT,DEBIT_CR_CD,CASH_ADV_TRXN_COMN_CNTRY_CD,MRCH_CATG_CD,MRCH_PSTL_CD,VISA_MAIL_PHN_CD,VISA_RQSTD_PMT_SVC_CD,MC_PMT_FACILITATOR_IDN_NUM
 100,0,2017-06-17,1537019,30.64,2017-06-15,D,CAN,5812,M2N5P5,,,0.0
@@ -943,7 +896,6 @@ def test_dupes_from_real_data(snowflake_session):
     compare_unq.report()
 
 
-@pytest.mark.snowflake
 def test_strings_with_joins_with_ignore_spaces(snowflake_session):
     df1 = snowflake_session.createDataFrame(
         [{"A": "HI", "B": " A"}, {"A": "BYE", "B": "A"}]
@@ -964,7 +916,6 @@ def test_strings_with_joins_with_ignore_spaces(snowflake_session):
     assert compare.intersect_rows_match()
 
 
-@pytest.mark.snowflake
 def test_decimal_with_joins_with_ignore_spaces(snowflake_session):
     df1 = snowflake_session.createDataFrame([{"A": 1, "B": " A"}, {"A": 2, "B": "A"}])
     df2 = snowflake_session.createDataFrame([{"A": 1, "B": "A"}, {"A": 2, "B": "A "}])
@@ -981,7 +932,6 @@ def test_decimal_with_joins_with_ignore_spaces(snowflake_session):
     assert compare.intersect_rows_match()
 
 
-@pytest.mark.snowflake
 def test_joins_with_ignore_spaces(snowflake_session):
     df1 = snowflake_session.createDataFrame([{"A": 1, "B": " A"}, {"A": 2, "B": "A"}])
     df2 = snowflake_session.createDataFrame([{"A": 1, "B": "A"}, {"A": 2, "B": "A "}])
@@ -993,7 +943,6 @@ def test_joins_with_ignore_spaces(snowflake_session):
     assert compare.intersect_rows_match()
 
 
-@pytest.mark.snowflake
 def test_joins_with_insensitive_lowercase_cols(snowflake_session):
     df1 = snowflake_session.createDataFrame([{"a": 1, "B": "A"}, {"a": 2, "B": "A"}])
     df2 = snowflake_session.createDataFrame([{"A": 1, "B": "A"}, {"A": 2, "B": "A"}])
@@ -1014,7 +963,6 @@ def test_joins_with_insensitive_lowercase_cols(snowflake_session):
     assert compare.intersect_rows_match()
 
 
-@pytest.mark.snowflake
 def test_joins_with_sensitive_lowercase_cols(snowflake_session):
     df1 = snowflake_session.createDataFrame(
         [{'"a"': 1, "B": "A"}, {'"a"': 2, "B": "A"}]
@@ -1030,7 +978,6 @@ def test_joins_with_sensitive_lowercase_cols(snowflake_session):
     assert compare.intersect_rows_match()
 
 
-@pytest.mark.snowflake
 def test_full_join_counts_all_matches(snowflake_session):
     df1 = snowflake_session.createDataFrame([{"A": 1, "B": 2}, {"A": 1, "B": 2}])
     df2 = snowflake_session.createDataFrame([{"A": 1, "B": 2}, {"A": 1, "B": 2}])
@@ -1040,7 +987,6 @@ def test_full_join_counts_all_matches(snowflake_session):
     assert compare.count_matching_rows() == 2
 
 
-@pytest.mark.snowflake
 def test_strings_with_ignore_spaces_and_join_columns(snowflake_session):
     df1 = snowflake_session.createDataFrame(
         [{"A": "HI", "B": "A"}, {"A": "BYE", "B": "A"}]
@@ -1062,7 +1008,6 @@ def test_strings_with_ignore_spaces_and_join_columns(snowflake_session):
     assert compare.count_matching_rows() == 2
 
 
-@pytest.mark.snowflake
 def test_integers_with_ignore_spaces_and_join_columns(snowflake_session):
     df1 = snowflake_session.createDataFrame([{"A": 1, "B": "A"}, {"A": 2, "B": "A"}])
     df2 = snowflake_session.createDataFrame([{"A": 1, "B": "A"}, {"A": 2, "B": "A"}])
@@ -1081,7 +1026,6 @@ def test_integers_with_ignore_spaces_and_join_columns(snowflake_session):
     assert compare.count_matching_rows() == 2
 
 
-@pytest.mark.snowflake
 def test_sample_mismatch(snowflake_session):
     data1 = """ACCT_ID,DOLLAR_AMT,NAME,FLOAT_FLD,DATE_FLD
     10000001234,123.45,George Maharis,14530.1555,2017-01-01
@@ -1119,7 +1063,6 @@ def test_sample_mismatch(snowflake_session):
     assert (output.NAME_DF1 != output.NAME_DF2).all()
 
 
-@pytest.mark.snowflake
 def test_all_mismatch_not_ignore_matching_cols_no_cols_matching(snowflake_session):
     data1 = """ACCT_ID,DOLLAR_AMT,NAME,FLOAT_FLD,DATE_FLD
     10000001234,123.45,George Maharis,14530.1555,2017-01-01
@@ -1159,7 +1102,6 @@ def test_all_mismatch_not_ignore_matching_cols_no_cols_matching(snowflake_sessio
     assert (~(output.DATE_FLD_DF1 != output.DATE_FLD_DF2)).values.sum() == 0
 
 
-@pytest.mark.snowflake
 def test_all_mismatch_not_ignore_matching_cols_some_cols_matching(snowflake_session):
     # Columns dollar_amt and name are matching
     data1 = """ACCT_ID,DOLLAR_AMT,NAME,FLOAT_FLD,DATE_FLD
@@ -1200,7 +1142,6 @@ def test_all_mismatch_not_ignore_matching_cols_some_cols_matching(snowflake_sess
     assert (~(output.DATE_FLD_DF1 != output.DATE_FLD_DF2)).values.sum() == 0
 
 
-@pytest.mark.snowflake
 def test_all_mismatch_ignore_matching_cols_some_cols_matching_diff_rows(
     snowflake_session,
 ):
@@ -1242,7 +1183,6 @@ def test_all_mismatch_ignore_matching_cols_some_cols_matching_diff_rows(
     assert not ("DOLLAR_AMT_DF1" in output and "DOLLAR_AMT_DF1" in output)
 
 
-@pytest.mark.snowflake
 def test_all_mismatch_ignore_matching_cols_some_cols_matching(snowflake_session):
     # Columns dollar_amt and name are matching
     data1 = """ACCT_ID,DOLLAR_AMT,NAME,FLOAT_FLD,DATE_FLD
@@ -1281,7 +1221,6 @@ def test_all_mismatch_ignore_matching_cols_some_cols_matching(snowflake_session)
     assert not ("DOLLAR_AMT_DF1" in output and "DOLLAR_AMT_DF1" in output)
 
 
-@pytest.mark.snowflake
 def test_all_mismatch_ignore_matching_cols_no_cols_matching(snowflake_session):
     data1 = """ACCT_ID,DOLLAR_AMT,NAME,FLOAT_FLD,DATE_FLD
         10000001234,123.45,George Maharis,14530.1555,2017-01-01
@@ -1321,7 +1260,6 @@ def test_all_mismatch_ignore_matching_cols_no_cols_matching(snowflake_session):
     assert (~(output.DATE_FLD_DF1 != output.DATE_FLD_DF2)).values.sum() == 0
 
 
-@pytest.mark.snowflake
 def test_all_mismatch_ignore_matching_cols_no_cols_matching_abs_tol_float(
     snowflake_session,
 ):
@@ -1363,7 +1301,6 @@ def test_all_mismatch_ignore_matching_cols_no_cols_matching_abs_tol_float(
     assert (~(output.DATE_FLD_DF1 != output.DATE_FLD_DF2)).values.sum() == 0
 
 
-@pytest.mark.snowflake
 def test_all_mismatch_ignore_matching_cols_no_cols_matching_abs_tol_dict(
     snowflake_session,
 ):
@@ -1409,7 +1346,6 @@ def test_all_mismatch_ignore_matching_cols_no_cols_matching_abs_tol_dict(
     assert (~(output.DATE_FLD_DF1 != output.DATE_FLD_DF2)).values.sum() == 1
 
 
-@pytest.mark.snowflake
 def test_all_mismatch_ignore_matching_cols_no_cols_matching_rel_tol_float(
     snowflake_session,
 ):
@@ -1451,7 +1387,6 @@ def test_all_mismatch_ignore_matching_cols_no_cols_matching_rel_tol_float(
     assert (~(output.DATE_FLD_DF1 != output.DATE_FLD_DF2)).values.sum() == 0
 
 
-@pytest.mark.snowflake
 def test_all_mismatch_ignore_matching_cols_no_cols_matching_rel_tol_dict(
     snowflake_session,
 ):
@@ -1508,7 +1443,6 @@ def test_all_mismatch_ignore_matching_cols_no_cols_matching_rel_tol_dict(
         ("INFINITY", np.inf),
     ],
 )
-@pytest.mark.snowflake
 def test_calculate_max_diff(snowflake_session, column, expected):
     pdf = pd.DataFrame(
         {
@@ -1533,7 +1467,6 @@ def test_calculate_max_diff(snowflake_session, column, expected):
     )
 
 
-@pytest.mark.snowflake
 def test_dupes_with_nulls_strings(snowflake_session):
     pdf1 = pd.DataFrame(
         {
@@ -1557,7 +1490,6 @@ def test_dupes_with_nulls_strings(snowflake_session):
     assert comp.subset()
 
 
-@pytest.mark.snowflake
 def test_dupes_with_nulls_ints(snowflake_session):
     pdf1 = pd.DataFrame(
         {
@@ -1581,7 +1513,6 @@ def test_dupes_with_nulls_ints(snowflake_session):
     assert comp.subset()
 
 
-@pytest.mark.snowflake
 def test_generate_id_within_group(snowflake_session):
     matrix = [
         (
@@ -1639,7 +1570,6 @@ def test_generate_id_within_group(snowflake_session):
         assert (actual["_TEMP_0"] == expected).all()
 
 
-@pytest.mark.snowflake
 def test_generate_id_within_group_single_join(snowflake_session):
     dataframe = snowflake_session.createDataFrame(
         [{"A": 1, "B": 2, "__INDEX": 1}, {"A": 1, "B": 2, "__INDEX": 2}]
@@ -1655,7 +1585,6 @@ def test_generate_id_within_group_single_join(snowflake_session):
 
 @mock.patch("datacompy.snowflake.render")
 @mock.patch("datacompy.snowflake.save_html_report")
-@pytest.mark.snowflake
 def test_save_html(mock_save_html, mock_render, snowflake_session):
     df1 = snowflake_session.createDataFrame([{"A": 1, "B": 2}, {"A": 1, "B": 2}])
     df2 = snowflake_session.createDataFrame([{"A": 1, "B": 2}, {"A": 1, "B": 2}])
@@ -1678,7 +1607,6 @@ def test_save_html(mock_save_html, mock_render, snowflake_session):
     assert args[1] == "test.html"  # The filename
 
 
-@pytest.mark.snowflake
 def test_full_join_counts_no_matches(snowflake_session):
     df1 = snowflake_session.createDataFrame([{"A": 1, "B": 2}, {"A": 1, "B": 3}])
     df2 = snowflake_session.createDataFrame([{"A": 1, "B": 4}, {"A": 1, "B": 5}])
@@ -1715,7 +1643,6 @@ def test_full_join_counts_no_matches(snowflake_session):
     )
 
 
-@pytest.mark.snowflake
 def test_full_join_counts_some_matches(snowflake_session):
     df1 = snowflake_session.createDataFrame([{"A": 1, "B": 2}, {"A": 1, "B": 3}])
     df2 = snowflake_session.createDataFrame([{"A": 1, "B": 2}, {"A": 1, "B": 5}])
@@ -1755,7 +1682,6 @@ def test_full_join_counts_some_matches(snowflake_session):
     )
 
 
-@pytest.mark.snowflake
 def test_non_full_join_counts_no_matches(snowflake_session):
     df1 = snowflake_session.createDataFrame(
         [{"A": 1, "B": 2, "C": 4}, {"A": 1, "B": 3, "C": 4}]
@@ -1803,7 +1729,6 @@ def test_non_full_join_counts_no_matches(snowflake_session):
     )
 
 
-@pytest.mark.snowflake
 def test_non_full_join_counts_some_matches(snowflake_session):
     df1 = snowflake_session.createDataFrame(
         [{"A": 1, "B": 2, "C": 4}, {"A": 1, "B": 3, "C": 4}]
@@ -1847,7 +1772,6 @@ def test_non_full_join_counts_some_matches(snowflake_session):
     )
 
 
-@pytest.mark.snowflake
 def test_custom_template_usage(snowflake_session):
     """Test using a custom template with template_path parameter."""
     df1 = snowflake_session.createDataFrame([("a", 1), ("b", 2)], ["id", "value"])
@@ -1884,7 +1808,6 @@ def test_custom_template_usage(snowflake_session):
             os.unlink(template_path)
 
 
-@pytest.mark.snowflake
 def test_template_without_extension(snowflake_session):
     """Test that template_path works without .j2 extension."""
     df1 = snowflake_session.createDataFrame([("a", 1), ("b", 2)], ["id", "value"])
@@ -1904,7 +1827,6 @@ def test_template_without_extension(snowflake_session):
             os.unlink(template_path)
 
 
-@pytest.mark.snowflake
 def test_nonexistent_template(snowflake_session):
     """Test that a clear error is raised when template file doesn't exist."""
     df1 = snowflake_session.createDataFrame([("a", 1), ("b", 2)], ["id", "value"])
@@ -1919,7 +1841,6 @@ def test_nonexistent_template(snowflake_session):
     ) or "nonexistent_template.j2" in str(exc_info.value)
 
 
-@pytest.mark.snowflake
 def test_template_context_variables(snowflake_session):
     """Test that all expected context variables are available in the template."""
     df1 = snowflake_session.createDataFrame([("a", 1), ("b", 2)], ["id", "value"])
@@ -1952,7 +1873,6 @@ def test_template_context_variables(snowflake_session):
 
 @mock.patch("datacompy.snowflake.save_html_report")
 @mock.patch("datacompy.snowflake.render")
-@pytest.mark.snowflake
 def test_html_report_generation(mock_render, mock_save_html, snowflake_session):
     """Test that HTML reports can be generated and saved to a file."""
     df1 = snowflake_session.createDataFrame([("a", 1), ("b", 2)], ["id", "value"])
